requirements.txt
# Web scraping
playwright==1.41.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2

